chore: refactor event-broker location
