+++ src/components/Button/Button.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import type {DOMProps, QAProps} from '../types';
 import {block} from '../utils/cn';
 import {isIcon} from '../utils/common';
-import {eventBroker} from '../utils/event-broker';
+import {eventBroker} from '../utils/eventBroker';
 import {isOfType} from '../utils/isOfType';
 
 import {ButtonIcon} from './ButtonIcon';
