+++ src/components/Link/Link.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import type {DOMProps, QAProps} from '../types';
 import {block} from '../utils/cn';
-import {eventBroker} from '../utils/event-broker';
+import {eventBroker} from '../utils/eventBroker';
 
 import './Link.scss';
