+++ src/components/utils/eventBroker/index.ts
@@ -1,0 +1,3 @@
+export {eventBroker} from './EventBroker';
+export {useEventBroker} from './useEventBroker';
+export type {EventBrokerSubscription, EventBrokerData} from './EventBroker';
