+++ src/components/utils/useCheckbox.ts
@@ -3,7 +3,7 @@ import React from 'react';
 import {useForkRef} from '../../hooks';
 import type {ControlProps} from '../types';
 
-import {eventBroker} from './event-broker';
+import {eventBroker} from './eventBroker';
 
 export function useCheckbox({
     name,
